Replace mysql-* packages with default-mysql-* (#6)
The mysql-server and mysql-client metapackages were removed in Debian
Buster. Install default-mysql-server and default-mysql-client instead.
They were already transitional in Stretch and caused mariadb packages to
be installed.

Using the "default"-prefixed metapackages will install the same packages
that the old transitional ones did in Stretch and ensure a new default
is followed if it changes in a future release. This change drops
compatibility with Jessie because the new packages don't exist yet in
that version.